requirements.txt
aiohttp==3.6.2
APScheduler==3.6.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cleverwrap==0.3.0.2
discord.py==1.3.3
idna==2.9
idna-ssl==1.1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
nose==1.3.7
pylint==2.4.4
pytz==2019.3
Random-Word==1.0.4
RandomWords==0.2.1
requests==2.23.0
six==1.14.0
soupsieve==1.9.5
tinydb==3.15.2
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==2.0.0
urllib3==1.25.8
websockets==8.1
wrapt==1.11.2
yapf==0.29.0
yarl==1.4.2
